requirements.txt
streamlit
matplotlib
Pillow
tensorflow
numpy
kagglehub